+++ Presentation/WindowsClient/Controls/Importer.cs
@@ -1,5 +1,5 @@
 ﻿using Rems.Infrastructure;
-using Rems.Infrastructure.Utilities;
+using Rems.Infrastructure.Excel;
 using System;
 using System.Data;
 using System.Drawing;
