+++ Presentation/WindowsClient/Models/Nodes/DataNode.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Drawing;
 using System.Windows.Forms;
-using Rems.Infrastructure;
+using Rems.Infrastructure.Excel;
 
 namespace WindowsClient.Models
 {
