+++ Presentation/WindowsClient/Models/Nodes/GroupNode.cs
@@ -1,6 +1,4 @@
-﻿using Rems.Application.Common;
-using System;
-using System.Data;
+﻿using System;
 
 namespace WindowsClient.Models
 {
