+++ Presentation/WindowsClient/Models/Nodes/InvalidNode.cs
@@ -1,4 +1,4 @@
-﻿using Rems.Infrastructure;
+﻿using Rems.Infrastructure.Excel;
 using System;
 using System.Data;
 using System.Windows.Forms;
